Return proper data type for InvoiceItem's amount and Transaction's amount fields which is supposed to be float not int data type

+++ src/app/Services/Transaction/RefundTransactionService.php
@@ -11,7 +11,7 @@ class RefundTransactionService
     {
     }
 
-    public function __invoke(Invoice $invoice): int
+    public function __invoke(Invoice $invoice): float
     {
         $paidAmount = $this->transactionRepository->sumOfPaidTransactions($invoice);
         $this->transactionRepository->refundSuccessfulTransactions($invoice);
